refactoring side nav module
