--- tobe_blog_web/src/admin/components/navigator/index.ts
@@ -1,3 +1,0 @@
-import SideNav from "./SideNav";
-
-export { SideNav };
